+++ translations/ksnip_de.ts
@@ -161,8 +161,16 @@ macht sie glatter.</translation>
         <translation>Aktualisieren</translation>
     </message>
     <message>
-        <source>Painter Settings</source>
-        <translation>Zeichner-Einstellungen</translation>
+        <source>Rotate Watermark</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>When enabled, Watermark will be added with a rotation of 45°</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Annotator Settings</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>