+++ translations/ksnip_gl.ts
@@ -157,7 +157,15 @@ make them more smooth.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Painter Settings</source>
+        <source>Rotate Watermark</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>When enabled, Watermark will be added with a rotation of 45°</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Annotator Settings</source>
         <translation type="unfinished"></translation>
     </message>
 </context>