+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * New: Add watermarks to annotated image. ([#199](https://github.com/DamirPorobic/kImageAnnotator/issues/199))
 * New: Add crop button to toolbar. ([#90](https://github.com/DamirPorobic/kImageAnnotator/issues/90))
 * New: Add undo and redo button on toolbar. ([#124](https://github.com/DamirPorobic/kImageAnnotator/issues/124))
+* New: Make if watermark is rotated a config option. ([#206](https://github.com/DamirPorobic/kImageAnnotator/issues/206))
 * New kImageAnnotator: Edit text box content. ([#51](https://github.com/DamirPorobic/kImageAnnotator/issues/51))
 * Changed: Change copy icon. ([#157](https://github.com/DamirPorobic/kImageAnnotator/issues/157))
 * Fixed: Translations not working for Windows and MacOS. ([#164](https://github.com/DamirPorobic/ksnip/issues/164))